+++ packages/job-worker/src/playout/take.ts
@@ -12,7 +12,7 @@ import { getCurrentTime } from '../lib'
 import { PartEndState, VTContent } from '@sofie-automation/blueprints-integration'
 import { DBPartInstance } from '@sofie-automation/corelib/dist/dataModel/PartInstance'
 import { ReadonlyDeep } from 'type-fest'
-import { getResolvedPieces } from './pieces'
+import { getResolvedPieces } from './resolvedPieces'
 import { clone, getRandomId, literal, stringifyError } from '@sofie-automation/corelib/dist/lib'
 import { protectString } from '@sofie-automation/corelib/dist/protectedString'
 import { updateTimeline } from './timeline/generate'
